Add linter rule to forbid unsafe code
Several security requirements will only be fulfilled by us implicitly (i.e. we don't have to add extra safeguards), if we never use unsafe rust.
As there is no need for unsafe in this project, we can just add this linter rule and have a bunch of requirements for free.

Signed-off-by: aziz.karabashov <aziz.karabashov@dsr-corporation.com>