Remove unicode escaping for forbidden characters
The current encoding solution defeats the purpose by applying an extra encoding to the forbidden characters, which can be easily detectable by a parser and automatically decoded.

The reason base64 encoding is specified in the ASDK-214 ticket, and the filtered version has no encoding (only alphanumeric and spaces allowed) is because it is unfeasible to automatically detect  and decode base64 on generic test.

This is a consequence of ADSK not assuming what type of parser will handle its' logs.

Signed-off-by: aziz.karabashov <aziz.karabashov@dsr-corporation.com>